backend/requirements.txt
fastapi>=0.105.0
uvicorn[standard]>=0.23.0
SQLAlchemy>=2.0.20
asyncpg>=0.27.0
databases>=0.7.0
pydantic>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0,>=3.2.0
python-dotenv>=1.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
